--- src/common/main/kotlin/tech/thatgravyboat/skyblockapi/api/events/party/DungeonPartyFinderQueueEvent.kt
@@ -1,11 +1,0 @@
-package tech.thatgravyboat.skyblockapi.api.events.party
-
-import tech.thatgravyboat.skyblockapi.api.area.dungeon.DungeonFloor
-import tech.thatgravyboat.skyblockapi.api.events.base.SkyBlockEvent
-
-class DungeonPartyFinderQueueEvent(
-    val floor: DungeonFloor,
-    val groupNote: String,
-    val dungeonLevelRequirement: Int,
-    val classLevelRequirement: Int
-) : SkyBlockEvent()
