+++ src/common/main/kotlin/tech/thatgravyboat/skyblockapi/api/events/party/PartyFinderQueueEvent.kt
@@ -1,7 +1,15 @@
 package tech.thatgravyboat.skyblockapi.api.events.party
 
+import tech.thatgravyboat.skyblockapi.api.area.dungeon.DungeonFloor
 import tech.thatgravyboat.skyblockapi.api.area.isle.kuudra.KuudraTier
 import tech.thatgravyboat.skyblockapi.api.events.base.SkyBlockEvent
+
+class DungeonPartyFinderQueueEvent(
+    val floor: DungeonFloor,
+    val groupNote: String,
+    val dungeonLevelRequirement: Int,
+    val classLevelRequirement: Int
+) : SkyBlockEvent()
 
 class KuudraPartyFinderQueueEvent(
     val tier: KuudraTier,
